Relocate DummyConnection for re-usability
